requirements.txt
# PDF Forensics Deep Hashing - Dependencies
# Python 3.8+ required

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Transformers for BERT
transformers>=4.30.0

# PDF Processing
PyMuPDF>=1.22.0

# Image Processing
Pillow>=9.5.0

# Data Handling
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning Utilities
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Progress Bars
tqdm>=4.65.0

# Optional: For GPU support (CUDA)
# Install PyTorch with CUDA from: https://pytorch.org/get-started/locally/

# Optional: For faster training
# nvidia-apex  # NVIDIA Apex for mixed precision (if available)

# Development Dependencies (optional)
# jupyter>=1.0.0
# ipython>=8.12.0
# pytest>=7.3.0